+++ src/replication/upstream.ts
@@ -27,7 +27,12 @@ import {
 } from './meta-instance';
 
 /**
- * Writes all document changes from the client to the master.
+ * Writes all document changes from the fork to the master.
+ * The upstream runs on two modes:
+ * - For inital replication, a checkpoint-iteration is used
+ * - For ongoing local writes, we just subscribe to the changeStream of the fork.
+ *   In contrast to the master, the fork can be assumed to never loose connection,
+ *   so we do not have to prepare for missed out events.
  */
 export function startReplicationUpstream<RxDocType>(
     state: RxStorageInstanceReplicationState<RxDocType>