+++ src/types/rx-storage.interface.d.ts
@@ -223,7 +223,15 @@ export interface RxStorageInstance<
      * and others error. We need this to have a similar behavior as most NoSQL databases.
      */
     bulkWrite(
-        documentWrites: BulkWriteRow<RxDocType>[]
+        documentWrites: BulkWriteRow<RxDocType>[],
+        /**
+         * Context will be used in all
+         * changeStream()-events that are emitted as a result
+         * of that bulkWrite() operation.
+         * Used in plugins so that we can detect that event X
+         * comes from operation Y.
+         */
+        context: string
     ): Promise<
         /**
          * returns the response, splitted into success and error lists.